refactor. Split rendering code per view
- render/render.js 코드를 뷰마다 나눠서 랜더링할 수 있도록 파일을 나눔
