+++ client/view/index.html
@@ -8,7 +8,7 @@
     <title>Vending Machine</title>
   </head>
   <body>
-    <script type="module" src="../js/render/render.js"></script>
+    <div class="wrap"></div>
     <script type="module" src="../js/main.js"></script>
   </body>
 </html>
